requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
jinja2==3.1.2
pyjwt==2.8.0
email-validator==2.0.0
tinydb==4.7.1